+++ src/converse-fullscreen.js
@@ -7,7 +7,7 @@ import "@converse/headless/converse-muc";
 import "converse-chatview";
 import "converse-controlbox";
 import "converse-singleton";
-import { converse } from "@converse/headless/converse-core";
+import { _converse, api, converse } from "@converse/headless/converse-core";
 import tpl_brand_heading from "templates/inverse_brand_heading.html";
 
 
@@ -26,14 +26,12 @@ converse.plugins.add('converse-fullscreen', {
 
         ControlBoxView: {
             createBrandHeadingHTML() {
-                const { _converse } = this.__super__;
                 return tpl_brand_heading({
                     'version_name': _converse.VERSION_NAME
                 });
             },
 
             insertBrandHeading () {
-                const { _converse } = this.__super__;
                 const el = _converse.root.getElementById('converse-login-panel');
                 el.parentNode.insertAdjacentHTML(
                     'afterbegin',
@@ -44,7 +42,7 @@ converse.plugins.add('converse-fullscreen', {
     },
 
     initialize () {
-        this._converse.api.settings.update({
+        api.settings.extend({
             chatview_avatar_height: 50,
             chatview_avatar_width: 50,
             hide_open_bookmarks: true,