+++ src/headless/converse-smacks.js
@@ -22,7 +22,7 @@ converse.plugins.add('converse-smacks', {
         // ====================================
         // Refer to docs/source/configuration.rst for explanations of these
         // configuration settings.
-        api.settings.update({
+        api.settings.extend({
             'enable_smacks': true,
             'smacks_max_unacked_stanzas': 5,
         });